+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 
 
+## [0.9.0] – 2022-11-18
+
+### New
+
+- `Client.getClientInfo`
+- `Trader.getIndexInfo`
+
 ## [0.8.0] – 2022-11-18
 
 ### New